[data-spike-sitrep] - Refactored a bit, things ended up both more readable and measurably more efficient (8% performance gain without trying) - TT
