+++ Sources/SwiftCurrent_IRGenerator/Extensions/URLExtensions.swift
@@ -1,0 +1,20 @@
+//
+//  URLExtensions.swift
+//  SwiftCurrent
+//
+//  Created by Tyler Thompson on 3/3/22.
+//  Copyright © 2022 WWT and Tyler Thompson. All rights reserved.
+//  
+
+import Foundation
+import ArgumentParser
+
+extension URL: ExpressibleByArgument {
+    public init?(argument: String) {
+        self.init(string: argument)
+    }
+
+    public var defaultValueDescription: String {
+        "A valid URL or valid source code"
+    }
+}
